Compile warnings

+++ src/pages/model_page.c
@@ -132,7 +132,7 @@ static void fixedid_cb(guiObject_t *obj, const void *data)
     if(Model.fixed_id == 0) {
         u32 id = PROTOCOL_CurrentID();
         if (id)
-            sprintf(mp->fixed_id, "%d", id);
+            sprintf(mp->fixed_id, "%d", (int)id);
         else
             mp->fixed_id[0] = 0;
     }
